recover base service & engine of weserver
